+++ src/app/app.tsx
@@ -15,13 +15,14 @@ import { WorkspaceView } from "./workspace/workspaceview";
 import { PluginsView } from "./pluginsview/pluginsview";
 import { BookmarksView } from "./bookmarks/bookmarks";
 import { HistoryView } from "./history/history";
+import { ConnectionsView } from "./connections/connections";
 import {
     ScreenSettingsModal,
     SessionSettingsModal,
     LineSettingsModal,
     ClientSettingsModal,
 } from "./common/modals/settings";
-import { RemotesModal } from "./connections/connections";
+import { RemotesModal } from "./connections_deprecated/connections";
 import { TosModal } from "./common/modals/modals";
 import { MainSideBar } from "./sidebar/sidebar";
 import {
@@ -30,6 +31,8 @@ import {
     AlertModal,
     AboutModal,
     CreateRemoteConnModal,
+    ViewRemoteConnDetailModal,
+    EditRemoteConnModal,
 } from "./common/modals/modals";
 import { ErrorBoundary } from "./common/error/errorboundary";
 import "./app.less";
@@ -85,14 +88,17 @@ class App extends React.Component<{}, {}> {
         let sessionSettingsModal = GlobalModel.sessionSettingsModal.get();
         let lineSettingsModal = GlobalModel.lineSettingsModal.get();
         let clientSettingsModal = GlobalModel.clientSettingsModal.get();
-        let remotesModel = GlobalModel.remotesModalModel;
-        let remotesModal = remotesModel.isOpen();
+        let remotesModel = GlobalModel.remotesModel;
+        let remotesModalMode = remotesModel.modalMode.get();
         let selectedRemoteId = remotesModel.selectedRemoteId.get();
+        let selectedRemote = GlobalModel.getRemote(selectedRemoteId);
+        let isAuthEditMode = remotesModel.isAuthEditMode();
         let remoteEdit = remotesModel.remoteEdit.get();
         let disconnected = !GlobalModel.ws.open.get() || !GlobalModel.waveSrvRunning.get();
         let hasClientStop = GlobalModel.getHasClientStop();
         let dcWait = this.dcWait.get();
         let platform = GlobalModel.getPlatform();
+
         if (disconnected || hasClientStop) {
             if (!dcWait) {
                 setTimeout(() => this.updateDcWait(true), 1500);
@@ -117,7 +123,6 @@ class App extends React.Component<{}, {}> {
         if (dcWait) {
             setTimeout(() => this.updateDcWait(false), 0);
         }
-        //console.log(`GlobalModel.activeMainView.get() = ${GlobalModel.activeMainView.get()}`); // @mike - if I remove this, I cant see plugins
         return (
             <div id="main" className={"platform-" + platform} onContextMenu={this.handleContextMenu}>
                 <div className="main-content">
@@ -127,6 +132,7 @@ class App extends React.Component<{}, {}> {
                         <WorkspaceView />
                         <HistoryView />
                         <BookmarksView />
+                        <ConnectionsView model={remotesModel} />
                     </ErrorBoundary>
                 </div>
                 <AlertModal />
@@ -136,8 +142,25 @@ class App extends React.Component<{}, {}> {
                 <If condition={GlobalModel.aboutModalOpen.get()}>
                     <AboutModal />
                 </If>
-                <If condition={remoteEdit !== null && !remoteEdit.old}>
+                <If condition={remoteEdit !== null && remotesModalMode === "add"}>
                     <CreateRemoteConnModal model={remotesModel} remoteEdit={remoteEdit} />
+                </If>
+                <If condition={selectedRemote != null}>
+                    <If condition={!isAuthEditMode && remotesModalMode === "read"}>
+                        <ViewRemoteConnDetailModal
+                            key={"remotedetail-" + selectedRemoteId}
+                            remote={selectedRemote}
+                            model={remotesModel}
+                        />
+                    </If>
+                    <If condition={remoteEdit !== null && isAuthEditMode && remotesModalMode === "edit"}>
+                        <EditRemoteConnModal
+                            key={"remotedetail-" + selectedRemoteId}
+                            remote={selectedRemote}
+                            model={remotesModel}
+                            remoteEdit={remoteEdit}
+                        />
+                    </If>
                 </If>
                 <If condition={screenSettingsModal != null}>
                     <ScreenSettingsModal
@@ -154,9 +177,6 @@ class App extends React.Component<{}, {}> {
                 </If>
                 <If condition={clientSettingsModal}>
                     <ClientSettingsModal />
-                </If>
-                <If condition={remotesModal}>
-                    <RemotesModal model={GlobalModel.remotesModalModel} />
                 </If>
             </div>
         );
