+++ src/app/common/modals/modals.less
@@ -460,7 +460,7 @@
             .action-buttons {
                 display: flex;
 
-                div.button {
+                button:first-child {
                     margin-right: 8px;
                 }
             }
@@ -468,29 +468,215 @@
     }
 }
 
-.wave-button {
-    display: flex;
-    padding: 6px 16px;
-    align-items: center;
-    gap: var(--sizing-2-xs, 4px);
-    border-radius: 6px;
-    height: auto;
+.wave-modal.rconndetail-modal {
+    .wave-modal-content.rconndetail-wave-modal-content {
+        width: 631px;
+        min-height: 565px;
+        overflow: visible;
+
+        .wave-modal-content-inner.rconndetail-wave-modal-content-inner {
+            display: flex;
+            padding-bottom: 0px;
+            flex-direction: column;
+            align-items: center;
+            gap: 20px;
+            flex-shrink: 0;
+
+            .rconndetail-wave-modal-body {
+                display: flex;
+                padding: 0px 20px;
+                align-items: flex-start;
+                width: 100%;
+                display: flex;
+                flex-direction: column;
+                gap: 16px;
+                align-self: stretch;
+
+                .name-header-actions-wrapper {
+                    display: flex;
+                    flex-direction: column;
+                    align-items: flex-start;
+                    gap: 12px;
+
+                    .rconndetail-name {
+                        color: @term-bright-white;
+                        font-size: 15px;
+                        font-weight: 500;
+                        line-height: 20px;
+                    }
+
+                    .header-actions {
+                        display: flex;
+                        justify-content: flex-end;
+                        align-items: flex-start;
+
+                        .wave-button {
+                            padding: 4px 15px;
+                            font-size: 11px;
+                            margin-right: 8px;
+                        }
+                    }
+                }
+
+                .remote-detail {
+                    .settings-field {
+                        display: flex;
+                        flex-direction: row;
+                        align-items: center;
+
+                        .settings-label {
+                            font-weight: bold;
+                            width: 12em;
+                            display: flex;
+                            flex-direction: row;
+                            align-items: center;
+                        }
+
+                        .settings-input {
+                            display: flex;
+                            flex-direction: row;
+                            align-items: center;
+                            color: @term-white;
+                        }
+                    }
+
+                    .settings-field:not(:first-child) {
+                        margin-top: 4px;
+                    }
+
+                    .status {
+                        display: flex;
+                        height: 30px;
+                        padding: 3px 8px;
+                        align-items: center;
+                        gap: 8px;
+                        align-self: stretch;
+                        border-radius: 6px;
+                        background: rgba(241, 246, 243, 0.08);
+                    }
+
+                    .terminal-wrapper {
+                        width: 100%;
+                        margin-top: 5px;
+
+                        .terminal-connectelem {
+                            height: 163px !important; // Needed to override plugin height
+
+                            .xterm-viewport {
+                                display: flex;
+                                padding: 6px 10px;
+                                gap: 8px;
+                                align-items: flex-start;
+                                align-self: stretch;
+                                border-radius: 6px;
+                                border: 1px solid var(--element-separator, rgba(241, 246, 243, 0.15));
+                                background: #080a08;
+                                height: 163px !important; // Needed to override plugin height
+                            }
+
+                            .xterm-screen {
+                                padding: 10px;
+                                width: 541px !important; // Needed to override plugin width
+                            }
+                        }
+                    }
+                }
+            }
+        }
+
+        .rconndetail-wave-modal-footer {
+            display: flex;
+            justify-content: flex-end;
+            width: 100%;
+            padding: 0 20px 20px;
+
+            .action-buttons {
+                display: flex;
+
+                button:first-child {
+                    margin-right: 8px;
+                }
+            }
+        }
+    }
 }
 
-.wave-button.color-green {
-    color: @term-bright-white;
-    background: @term-green !important; // !important is needed to override the default button color
-    box-shadow: 0px 1px 3px 0px rgba(0, 0, 0, 0.4), 0px 0px 0.5px 0px rgba(0, 0, 0, 0.5),
-        0px 0px 0.5px 0px rgba(255, 255, 255, 0.8) inset, 0px 0.5px 0px 0px rgba(255, 255, 255, 0.6) inset;
+.wave-modal.erconn-modal {
+    .wave-modal-content.erconn-wave-modal-content {
+        width: 502px;
+        min-height: 411px;
+        overflow: visible;
 
-    &:hover {
-        color: @term-bright-white;
+        .wave-modal-content-inner.erconn-wave-modal-content-inner {
+            display: flex;
+            padding-bottom: 0px;
+            flex-direction: column;
+            align-items: center;
+            gap: 20px;
+            flex-shrink: 0;
+
+            .erconn-wave-modal-body {
+                display: flex;
+                padding: 0px 20px;
+                flex-direction: column;
+                align-items: flex-start;
+                gap: 12px;
+                align-self: stretch;
+                width: 100%;
+
+                > div {
+                    width: 100%;
+                }
+
+                .name-actions-section {
+                    margin-bottom: 10px;
+                    display: flex;
+                    flex-direction: column;
+                    align-items: flex-start;
+                    gap: 12px;
+
+                    .name {
+                        color: @term-bright-white;
+                        font-size: 15px;
+                        font-weight: 500;
+                        line-height: 20px;
+                    }
+
+                    .header-actions {
+                        display: flex;
+                        justify-content: flex-end;
+                        align-items: flex-start;
+
+                        .wave-button {
+                            padding: 4px 15px;
+                            font-size: 11px;
+                            margin-right: 8px;
+                        }
+                    }
+                }
+            }
+        }
+
+        .erconn-wave-modal-footer {
+            display: flex;
+            justify-content: flex-end;
+            width: 100%;
+            padding: 0 20px 20px;
+
+            .action-buttons {
+                display: flex;
+
+                button:first-child {
+                    margin-right: 8px;
+                }
+            }
+        }
     }
 }
 
 .wave-button.color-standard {
     color: @term-white;
-    background: var(--overlays-white-6, rgba(255, 255, 255, 0.12));
+    background: rgba(255, 255, 255, 0.12);
 
     &:hover {
         color: @term-white;
